move files around & add run config to git
