+++ backend/src/main/java/com/filegenie/backend/Application.java
@@ -1,4 +1,4 @@
-package com.filegenie.filegeniebackend;
+package com.filegenie.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
