+++ backend/src/main/java/com/filegenie/backend/Entities/Field.java
@@ -1,11 +1,11 @@
-package com.filegenie.filegeniebackend.Entities;
+package com.filegenie.backend.Entities;
 
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.util.ArrayList;
+
 import java.util.List;
 import lombok.Builder;
