+++ backend/src/main/java/com/filegenie/backend/Repositories/FieldRepository.java
@@ -1,6 +1,6 @@
-package com.filegenie.filegeniebackend.Repositories;
+package com.filegenie.backend.Repositories;
 
-import com.filegenie.filegeniebackend.Entities.Field;
+import com.filegenie.backend.Entities.Field;
 import org.springframework.data.repository.CrudRepository;
 
 public interface FieldRepository extends CrudRepository<Field, Integer> {
