+++ backend/src/main/java/com/filegenie/backend/Repositories/UserRepository.java
@@ -1,6 +1,6 @@
-package com.filegenie.filegeniebackend.Repositories;
+package com.filegenie.backend.Repositories;
 
-import com.filegenie.filegeniebackend.Entities.User;
+import com.filegenie.backend.Entities.User;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository extends CrudRepository<User, Integer> {
