+++ backend/src/main/java/com/filegenie/backend/Repositories/ValueRepository.java
@@ -1,6 +1,6 @@
-package com.filegenie.filegeniebackend.Repositories;
+package com.filegenie.backend.Repositories;
 
-import com.filegenie.filegeniebackend.Entities.Value;
+import com.filegenie.backend.Entities.Value;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ValueRepository extends CrudRepository<Value, Integer> {}
