+++ backend/src/test/java/com/filegenie/backend/FileGenieBackEndApplicationTests.java
@@ -1,4 +1,4 @@
-package com.filegenie.filegeniebackend;
+package com.filegenie.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
